requirements.txt
autopep8==1.4.4
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-blog-zinnia==0.20
django-contrib-comments==1.9.1
django-dotenv==1.4.1
django-el-pagination==3.2.4
django-js-asset==1.2.2
django-modelcluster==5.0
django-mptt==0.10.0
django-tagging==0.4.6
django-taggit==1.2.0
django-treebeard==4.3
django-xmlrpc==0.1.8
djangorestframework==3.10.3
draftjs-exporter==2.1.7
elasticsearch==7.1.0
html5lib==1.0.1
idna==2.8
mots-vides==2015.5.11
Pillow==6.2.1
pycodestyle==2.5.0
pyparsing==2.4.5
pytz==2019.3
regex==2019.11.1
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.7
wagtailfontawesome==1.1.4
webencodings==0.5.1
Willow==1.3
